Updating the recipe to change membership level on expiration only
The previous recipe was no longer working to change the level on expiration. If the member canceled their membership on the old recipe, the level was canceled immediately. With this updated recipe, a user canceling a subscription will get an expiration date, and when it expires, they are downgraded to the set free level.