Create send-admin-change-email-after-add-member.php
This updated snippet is part of a rewrite of the following article:
https://www.paidmembershipspro.com/send-signup-emails-to-zapier-added-pmpro-members/

The original snippet would cause the new user notification to be sent twice. This updated snippet drops the call to `wp_new_user_notification()` as recent versions of the Zapier Add On already trigger this email when a new member is added.

The Zapier Add On now includes a `pmproz_after_add_member` hook that can be used to perform actions when a member is added via Zapier. This snippet uses this hook instead of `pmpro_after_change_membership_level` to send the admin change email.